requirements.txt
pandas
numpy
requests
psycopg2-binary
earthengine-api
